src/data/services/jwt_service.py
from jose import jwt


class JWTService:
    def __init__(self, secret_key: str, algorithm: str):
        self.secret_key = secret_key
        self.algorithm = algorithm

    def generate_login_token(self, email: str) -> str:
        to_encode = {
            "email": email,
        }

        encoded_jwt = jwt.encode(
            to_encode,
            self.secret_key,
            algorithm=self.algorithm
        )

        return encoded_jwt

    def verify_login_token(self, token: str) -> str | None:
        try:
            payload = jwt.decode(token, self.secret_key, algorithms=[self.algorithm])
            email: str = payload.get("email")

            if email is None:
                return None
            else:
                return email

        except Exception as e:
            return None
