requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
python-multipart==0.0.21
python-dotenv==1.2.1
pydantic-settings==2.12.0
sqlalchemy==2.1.0b1
pydantic==2.12.5
python-jose[cryptography]
